+++ NEWS.md
@@ -1,6 +1,7 @@
 billboarder (unreleased)
 ==================
 
+* Update of the billboard.js library to 3.4.0.
 * Update of the billboard.js library to 3.2.0, see the [release note](https://netil.medium.com/billboard-js-3-2-release-sparkline-tableview-plugins-more-32217ddc869a)
 * Update of the billboard.js library to 3.1.1, see the [release note](https://netil.medium.com/billboard-js-3-1-release-subchart-data-loading-enhancements-e98be5eebdcb)
 * Fixed bug with export outside shiny.