+++ NEWS.md
@@ -1,6 +1,7 @@
-billboarder (unreleased)
+billboarder 0.4.0
 ==================
 
+* Update of the billboard.js library to 3.6.3, see the [release note](https://netil.medium.com/billboard-js-3-6-release-official-react-wrapper-new-enhancements-2dbf1ffc4d1c)
 * Update of the billboard.js library to 3.5.1, see the [release note](https://netil.medium.com/billboard-js-3-5-release-new-boost-options-7c951e576076)
 * Update of the billboard.js library to 3.4.0, see the [release note](https://netil.medium.com/billboard-js-3-4-release-new-polar-chart-83d604c8706c)
 * Update of the billboard.js library to 3.2.0, see the [release note](https://netil.medium.com/billboard-js-3-2-release-sparkline-tableview-plugins-more-32217ddc869a)